+++ eras/babbage/impl/test/Main.hs
@@ -3,6 +3,7 @@
 module Main where
 
 import Cardano.Ledger.Babbage (BabbageEra)
+import Data.Proxy (Proxy (..))
 import qualified Test.Cardano.Ledger.Alonzo.Binary.CostModelsSpec as CostModelsSpec
 import qualified Test.Cardano.Ledger.Alonzo.Binary.TxWitsSpec as TxWitsSpec
 import qualified Test.Cardano.Ledger.Babbage.Binary.CddlSpec as CddlSpec
@@ -11,6 +12,7 @@ import qualified Test.Cardano.Ledger.Babbage.GoldenSpec as Golden
 import qualified Test.Cardano.Ledger.Babbage.GoldenTranslation as GoldenTranslation
 import qualified Test.Cardano.Ledger.Babbage.Imp as Imp
 import Test.Cardano.Ledger.Babbage.ImpTest ()
+import qualified Test.Cardano.Ledger.Babbage.TxInfoSpec as TxInfo
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Core.JSON (roundTripJsonEraSpec)
 import Test.Cardano.Ledger.Shelley.JSON (roundTripJsonShelleyEraSpec)
@@ -19,6 +21,7 @@ main :: IO ()
 main =
   ledgerTestMain $
     describe "Babbage" $ do
+      TxInfo.spec (Proxy @BabbageEra)
       GoldenTranslation.spec
       Golden.spec
       BinarySpec.spec