+++ eras/shelley/impl/CHANGELOG.md
@@ -62,6 +62,7 @@
 
 ### `testlib`
 
+* Added `Examples` with: `LedgerExamples`, `ledgerExamples`, `mkLedgerExamples`, `exampleCerts`,`exampleWithdrawals`, `exampleAuxDataMap`, `exampleNonMyopicRewards`, `exampleCoin`, `examplePayKey`, `exampleStakeKey`, `exampleNewEpochState`, `examplePoolDistr`, `examplePoolParams`, `exampleTxIns`, `exampleProposedPPUpdates`, `exampleByronAddress`, `testShelleyGenesis`, `keyToCredential`, `mkDSIGNKeyPair`, `mkKeyHash`, `mkScriptHash`, `mkWitnessesPreAlonzo`, `seedFromByte`, `seedFromWords`
 * Deprecated `getRewardAccountAmount`, `lookupReward` and `getReward`
 * Add `getBalance`, `lookupBalance`, `getAccountBalance`, `lookupAccountBalance`
 * Add `mkShelleyTestAccountState`, `shelleyAccountsFromAccountsMap`, `shelleyAccountsToUMap`